+++ src/Simulation/EntryPointDriver/Mock/NoOpQirSubmitter.cs
@@ -1,10 +1,12 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
 using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
 using Microsoft.Quantum.Runtime;
 
-namespace Microsoft.Quantum.EntryPointDriver.Mocks
+namespace Microsoft.Quantum.EntryPointDriver.Mock
 {
     /// <summary>
     /// A QIR submitter that does nothing.
