+++ src/Simulation/EntryPointDriver/Mock/NoOpQuantumMachine.cs
@@ -1,25 +1,24 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-using Microsoft.Azure.Quantum.Exceptions;
 using Microsoft.Quantum.Runtime;
 using Microsoft.Quantum.Simulation.Core;
 using System;
 using System.Threading.Tasks;
 
-namespace Microsoft.Quantum.EntryPointDriver.Mocks
+namespace Microsoft.Quantum.EntryPointDriver.Mock
 {
     /// <summary>
-    /// A quantum machine that always has an error.
+    /// A quantum machine that does nothing.
     /// </summary>
-    internal class ErrorQuantumMachine : IQuantumMachine
+    internal class NoOpQuantumMachine : IQuantumMachine
     {
         /// <summary>
-        /// The target ID for the error quantum machine.
+        /// The target ID for the no-op quantum machine.
         /// </summary>
-        internal const string TargetId = "test.error";
+        internal const string TargetId = "test.noop";
 
-        public string ProviderId => nameof(ErrorQuantumMachine);
+        public string ProviderId => nameof(NoOpQuantumMachine);
 
         public string Target => TargetId;
 
@@ -35,7 +34,7 @@ namespace Microsoft.Quantum.EntryPointDriver.Mocks
             EntryPointInfo<TInput, TOutput> info,
             TInput input,
             IQuantumMachineSubmissionContext submissionContext,
-            IQuantumMachine.ConfigureJob configureJobCallback) =>
+            IQuantumMachine.ConfigureJob configureJobCallback) => 
             throw new NotSupportedException();
 
         public Task<IQuantumMachineOutput<TOutput>> ExecuteAsync<TInput, TOutput>(
@@ -45,7 +44,7 @@ namespace Microsoft.Quantum.EntryPointDriver.Mocks
         public Task<IQuantumMachineOutput<TOutput>> ExecuteAsync<TInput, TOutput>(
             EntryPointInfo<TInput, TOutput> info,
             TInput input,
-            IQuantumMachineExecutionContext executionContext,
+            IQuantumMachineExecutionContext executionContext, 
             IQuantumMachine.ConfigureJob configureJobCallback) =>
             throw new NotSupportedException();
 
@@ -66,21 +65,21 @@ namespace Microsoft.Quantum.EntryPointDriver.Mocks
 
         public Task<IQuantumMachineJob> SubmitAsync<TInput, TOutput>(
             EntryPointInfo<TInput, TOutput> info, TInput input) =>
-            throw new AzureQuantumException("This quantum machine always has an error.");
+            Task.FromResult<IQuantumMachineJob>(new ExampleJob());
 
         public Task<IQuantumMachineJob> SubmitAsync<TInput, TOutput>(
-            EntryPointInfo<TInput, TOutput> info, TInput input, IQuantumMachineSubmissionContext submissionContext) => 
+            EntryPointInfo<TInput, TOutput> info, TInput input, IQuantumMachineSubmissionContext submissionContext) =>
             SubmitAsync(info, input);
 
         public Task<IQuantumMachineJob> SubmitAsync<TInput, TOutput>(
             EntryPointInfo<TInput, TOutput> info,
-            TInput input,
-            IQuantumMachineSubmissionContext submissionContext,
+            TInput input, 
+            IQuantumMachineSubmissionContext submissionContext, 
             IQuantumMachine.ConfigureJob configureJobCallback) =>
             SubmitAsync(info, input);
 
         public (bool IsValid, string Message) Validate<TInput, TOutput>(
             EntryPointInfo<TInput, TOutput> info, TInput input) =>
-            (false, "This quantum machine always has an error.");
+            (true, string.Empty);
     }
 }
